Setup environment for running ADSimDectector
This environment is based on a container with all modules prebuilt.
Submodules in the repository (ADCore for now) are mapped into the
container overwriting prebuilt ones. This allows to easily include the
changed module into Area Detector modules without needing to rebuild
everything.

Since the changes are stored in the mapped volume (ie. in the host
machine), multiple execution of the container without any changes won't
need recompiling the source code.