Add configurable start up script
This includes the configuration of the IOC and the area detector
plugins. These files have been based on the ADSimDetector st_base.cmd
and ADCore EXAMPLE_commonPlugins.cmd. Note that several plugins have
been removed from the IOC so that we can focus on having the performance
analysis on what we need.